recipe/requirements.txt
asgiref==3.8.1
Django==5.0.4
uvicorn==0.34.0
django-cors-headers==4.3.1
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.11.0
pre-commit==4.0.1
mysqlclient==2.2.6
django-prometheus==2.3.1
PyYAML>=6.0
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
pytest==8.3.4
requests==2.32.3
pytest-django==4.10.0